Round temperature for forecasts.

+++ api/src/main/java/dev/andrevv/weather/api/entity/ForecastItem.java
@@ -7,5 +7,5 @@ import java.time.LocalDate;
 @Data
 public class ForecastItem {
     private final LocalDate date;
-    private final double temperature;
+    private final long temperature;
 }
